+++ docs/CHANGELOG.md
@@ -78,6 +78,7 @@ include_toc: true
     * `Send Inventory To Server` (Enabled By Default)
     * `Clear Old Chat Messages When Joining Worlds` (Enabled By Default)
     * `Improve API Chat Messages` (Enabled By Default)
+    * `Sync Chunks To Disk After Saving` (Enabled By Default)
   * New Bug Fixes (All Enabled By Default)
     * `Fix Held Item Caching`
     * `Fix Pigmen Burning In The Sun`